feat: Handle API for Seat Type

+++ src/api/v1/seat-type/seat-type.service.ts
@@ -1,4 +1,9 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateSeatTypeDto } from './dto/create-seat-type.dto';
 import { UpdateSeatTypeDto } from './dto/update-seat-type.dto';
 import { PrismaService } from 'src/prisma.service';
@@ -8,45 +13,67 @@ export class SeatTypeService {
   constructor(private prisma: PrismaService) {}
   async createSeat(createSeatTypeDto: CreateSeatTypeDto) {
     try {
-      return await this.prisma.seatType.create({ data: createSeatTypeDto });
+      const seat = await this.prisma.seatType.create({
+        data: createSeatTypeDto,
+      });
+      return seat;
     } catch (err) {
-      console.error(err);
       throw new HttpException('Create fail!', HttpStatus.BAD_REQUEST);
     }
   }
 
   async findAllSeat() {
     try {
-      return await this.prisma.seatType.findMany();
+      const allSeats = await this.prisma.seatType.findMany();
+      return allSeats;
     } catch (err) {
-      throw new HttpException('Something wrong happend!', HttpStatus.CONFLICT);
+      throw new HttpException(
+        'Something wrong happend!',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
     }
   }
 
   async findOne(id: number) {
     try {
-      return await this.prisma.seatType.findUnique({ where: { id } });
+      const seat = await this.prisma.seatType.findUnique({ where: { id } });
+      return seat;
     } catch (err) {
-      throw new HttpException('Something wrong happend!', HttpStatus.CONFLICT);
+      throw new HttpException(
+        'Something wrong happend!',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
     }
   }
 
   async updateSeat(id: number, updateSeatTypeDto: UpdateSeatTypeDto) {
     try {
-      return await this.prisma.seatType.update({
+      const seatType = await this.prisma.seatType.findUnique({
+        where: { id },
+      });
+      if (!seatType) throw new NotFoundException('Seat Type Not Found!');
+      const updatedSeat = await this.prisma.seatType.update({
         where: { id },
         data: updateSeatTypeDto,
       });
+      return updatedSeat;
     } catch (err) {
-      throw new HttpException('Something wrong happend!', HttpStatus.CONFLICT);
+      throw new HttpException('Update failed!', HttpStatus.CONFLICT);
     }
   }
 
   async removeSeat(id: number) {
     try {
-      return await this.prisma.seatType.delete({ where: { id } });
+      const seatType = await this.prisma.seatType.findUnique({ where: { id } });
+      if (!seatType) throw new NotFoundException('Seat Type Not Found!');
+      await this.prisma.seat.deleteMany({
+        where: {
+          seatTypeId: id,
+        },
+      });
+      await this.prisma.seatType.delete({ where: { id } });
     } catch (err) {
-      throw new HttpException('Something wrong happend!', HttpStatus.CONFLICT);
+      throw new HttpException('Delete failed', HttpStatus.CONFLICT);
     }
   }
 }
